+++ lib/routes.dart
@@ -32,8 +32,8 @@ import 'package:picos/screens/questionaire_screen/questionaire_screen.dart';
 import 'package:picos/screens/study_nurse_screen/menu_screen/edit_patient_screen.dart';
 import 'package:picos/screens/study_nurse_screen/menu_screen/menu_main_screen.dart';
 import 'package:picos/screens/study_nurse_screen/catalog_of_items_screen/catalog_of_items_screen.dart';
-import 'package:picos/screens/visits_screen/my_visits_screen.dart';
-import 'package:picos/screens/visits_screen/visits_screen.dart';
+import 'package:picos/screens/my_visits_screen/my_visits_screen.dart';
+import 'package:picos/screens/my_visits_screen/add_visit_screen.dart';
 import 'package:picos/screens/study_nurse_screen/configuration_screen/configuration_finished_screen.dart';
 import 'package:picos/screens/study_nurse_screen/configuration_screen/configuration_pages.dart';
 import 'package:picos/screens/study_nurse_screen/configuration_screen/configuration_screen.dart';
@@ -76,7 +76,7 @@ class Routes {
       '/legals-screen': (BuildContext ctx) => const LegalsScreen(),
       '/privacy-notice-screen': (BuildContext ctx) =>
           const PrivacyNoticeScreen(),
-      '/visits-screen/visits': (BuildContext ctx) => const VisitsScreen(),
+      '/visits-screen/visits': (BuildContext ctx) => const AddVisitScreen(),
       '/visits-screen/my-visits-screen':(BuildContext ctx) => const MyVisitsScreen(),
       '/login-screen': (BuildContext ctx) => const LoginScreen(),
       '/questionnaire-screen/questionnaire-screen': (BuildContext ctx) =>
