+++ lib/screens/my_visits_screen/visit_item.dart
